add function - judge contract type page to display
