+++ src/components/screens/paidMemberMng/matterForm/create/SelectContractTypeForm/types.ts
@@ -1,11 +1,4 @@
-export const SelectContractType = {
-  ChiraCeo: "チラCEO",
-  JoiningAgency: "代理店加盟",
-  LetterMeasuresAndChiraCeo: "手紙施策 x チラCEO",
-  LetterGet: "レターゲット",
-} as const;
-
-export type SelectContractKeyType = keyof typeof SelectContractType;
+import { SelectContractKeyType } from "../types";
 
 export type FormData = {
   selectContractType: SelectContractKeyType;
