+++ src/components/screens/paidMemberMng/matterForm/create/types.ts
@@ -1,0 +1,8 @@
+export const SelectContractType = {
+  ChiraCeo: "チラCEO",
+  JoiningAgency: "代理店加盟",
+  LetterMeasuresAndChiraCeo: "手紙施策 x チラCEO",
+  LetterGet: "レターゲット",
+} as const;
+
+export type SelectContractKeyType = keyof typeof SelectContractType;
